refactor(contentful-personalization): isolate live-debug subskill [NT-0000]
