+++ src/skills/contentful-personalization/skill.test.ts
@@ -5,6 +5,7 @@ import skill from './skill.js';
 import { derivePackagesToInstall } from './actions/install-packages.js';
 import doctorSkill from './subskills/doctor.js';
 import developSkill from './subskills/develop.js';
+import liveDebugSkill from './subskills/live-debug.js';
 
 // --- Dispatcher routing tests ---
 
@@ -107,16 +108,18 @@ test('classify routes live URL requests to live-debug', async () => {
         requestedUrl: 'https://example.com/personalized',
         reasoning: 'User explicitly asked to inspect a live URL',
       },
-      'live-debug-check-mcp': {
+      'live-debug/check-mcp': {
         mcpAvailable: false,
         matchedTools: [],
       },
-      'live-debug-install-mcp': { message: 'Install Chrome DevTools MCP and rerun' },
+      'live-debug/install-mcp': { message: 'Install Chrome DevTools MCP and rerun' },
     }),
   });
 
-  assert.ok(result.path.includes('live-debug-check-mcp'));
-  assert.ok(result.path.includes('live-debug-install-mcp'));
+  assert.equal(result.redirectedTo?.kind, 'subskill');
+  assert.equal(result.redirectedTo?.name, 'live-debug');
+  assert.ok(result.path.includes('live-debug/check-mcp'));
+  assert.ok(result.path.includes('live-debug/install-mcp'));
   assert.ok(!result.path.includes('doctor/explore'));
 });
 
@@ -176,20 +179,15 @@ test('reference without topic routes to pick-topic', async () => {
 });
 
 test('live-debug uses provided URL and finishes when runtime looks healthy', async () => {
-  const result = await runComposite(skill, {
+  const result = await runSkill(liveDebugSkill, {
+    params: { requestedUrl: 'https://example.com/personalized' },
     host: { toolsAvailable: ['mcp__chrome-devtools__new_page', 'mcp__chrome-devtools__list_network_requests'] },
     model: mockModel({
-      classify: {
-        intent: 'live-debug',
-        confidence: 0.96,
-        requestedUrl: 'https://example.com/personalized',
-        reasoning: 'User wants runtime verification for a live page',
-      },
-      'live-debug-check-mcp': {
+      'check-mcp': {
         mcpAvailable: true,
         matchedTools: ['mcp__chrome-devtools__new_page', 'mcp__chrome-devtools__list_network_requests'],
       },
-      'live-debug-inspect': {
+      inspect: {
         url: 'https://example.com/personalized',
         overallStatus: 'pass',
         summary: 'Runtime behavior looks healthy.',
@@ -207,32 +205,24 @@ test('live-debug uses provided URL and finishes when runtime looks healthy', asy
         recommendations: [],
         shouldRunDoctor: false,
       },
-      'live-debug-report': { message: 'Looks healthy' },
+      report: { message: 'Looks healthy' },
     }),
   });
 
-  assert.ok(result.path.includes('live-debug-check-mcp'));
-  assert.ok(result.path.includes('live-debug-inspect'));
-  assert.ok(result.path.includes('live-debug-report'));
-  assert.ok(!result.path.includes('live-debug-request-url'));
-  assert.ok(!result.path.includes('doctor/explore'));
+  assert.deepEqual(result.path, ['check-mcp', 'inspect', 'report']);
+  assert.ok(!result.path.includes('request-url'));
 });
 
 test('live-debug asks for URL when one was not provided', async () => {
-  const result = await runComposite(skill, {
+  const result = await runSkill(liveDebugSkill, {
     host: { toolsAvailable: ['mcp__chrome-devtools__new_page'] },
     model: mockModel({
-      classify: {
-        intent: 'live-debug',
-        confidence: 0.9,
-        reasoning: 'User asked for live browser debugging but did not include a URL',
-      },
-      'live-debug-check-mcp': {
+      'check-mcp': {
         mcpAvailable: true,
         matchedTools: ['mcp__chrome-devtools__new_page'],
       },
-      'live-debug-request-url': { url: 'https://example.com/live' },
-      'live-debug-inspect': {
+      'request-url': { url: 'https://example.com/live' },
+      inspect: {
         url: 'https://example.com/live',
         overallStatus: 'pass',
         summary: 'Runtime behavior looks healthy.',
@@ -243,29 +233,23 @@ test('live-debug asks for URL when one was not provided', async () => {
         recommendations: [{ priority: 'info', message: 'Retry the page with known personalized content if you expected network activity.', category: 'runtime' }],
         shouldRunDoctor: false,
       },
-      'live-debug-report': { message: 'Done' },
+      report: { message: 'Done' },
     }),
   });
 
-  assert.ok(result.path.includes('live-debug-request-url'));
-  assert.ok(result.path.includes('live-debug-inspect'));
+  assert.deepEqual(result.path, ['check-mcp', 'request-url', 'inspect', 'report']);
 });
 
-test('live-debug escalates into doctor when runtime looks suspicious', async () => {
-  const result = await runComposite(skill, {
+test('live-debug recommends doctor when runtime looks suspicious', async () => {
+  const result = await runSkill(liveDebugSkill, {
+    params: { requestedUrl: 'https://example.com/personalized' },
     host: { toolsAvailable: ['mcp__chrome-devtools__new_page', 'mcp__chrome-devtools__list_console_messages'] },
     model: mockModel({
-      classify: {
-        intent: 'live-debug',
-        confidence: 0.98,
-        requestedUrl: 'https://example.com/personalized',
-        reasoning: 'User wants to inspect a live page and the runtime looks suspect',
-      },
-      'live-debug-check-mcp': {
+      'check-mcp': {
         mcpAvailable: true,
         matchedTools: ['mcp__chrome-devtools__new_page', 'mcp__chrome-devtools__list_console_messages'],
       },
-      'live-debug-inspect': {
+      inspect: {
         url: 'https://example.com/personalized',
         overallStatus: 'warn',
         summary: 'The page showed runtime symptoms that look like a setup issue.',
@@ -288,27 +272,11 @@ test('live-debug escalates into doctor when runtime looks suspicious', async () 
         ],
         shouldRunDoctor: true,
       },
-      'live-debug-report': { message: 'Escalating to doctor' },
-      'doctor/explore': {
-        framework: 'nextjs-app',
-        projectPath: '.',
-        explorationSummary: 'Provider setup looks incomplete.',
-        concerns: ['Provider not found'],
-      },
-      'doctor/scan-credentials': {
-        envVars: [{ name: 'NINETAILED_API_KEY', status: 'missing' }],
-      },
-      'doctor/confirm-credentials': { hasCredentials: false },
-      'doctor/review': { overallStatus: 'fail', recommendations: [], summary: 'Static diagnosis found issues.' },
-      'doctor/report': { choice: 'no' },
-      'doctor/done': { message: 'Done' },
+      report: { message: 'Run doctor next' },
     }),
   });
 
-  assert.ok(result.path.includes('live-debug-report'));
-  assert.ok(result.path.includes('doctor/explore'));
-  assert.equal(result.redirectedTo?.kind, 'subskill');
-  assert.equal(result.redirectedTo?.name, 'doctor');
+  assert.deepEqual(result.path, ['check-mcp', 'inspect', 'report']);
 });
 
 // --- Doctor sub-skill tests ---
