+++ src/skills/contentful-personalization/skill.ts
@@ -1,83 +1,9 @@
-import { skill, type, prompt, act, render, view, terminal } from '@contentful/skill-kit';
+import { skill, type, prompt, act } from '@contentful/skill-kit';
 import onboardSkill from './subskills/onboard.js';
+import liveDebugSkill from './subskills/live-debug.js';
 import doctorSkill from './subskills/doctor.js';
 import developSkill from './subskills/develop.js';
-import { RuntimeCheckResult } from './schemas.js';
 import { VERSION } from './version.js';
-
-function getChromeDevToolsToolMatches(tools: string[]) {
-  return tools.filter((tool) => tool.startsWith('mcp__chrome-devtools__') || tool.includes('chrome-devtools'));
-}
-
-function getLiveDebugUrl(store: {
-  steps: {
-    classify?: { requestedUrl?: string };
-    'gather-context'?: { requestedUrl?: string };
-    'live-debug-request-url'?: { url?: string };
-  };
-}) {
-  return (
-    store.steps['live-debug-request-url']?.url ??
-    store.steps.classify?.requestedUrl ??
-    store.steps['gather-context']?.requestedUrl ??
-    ''
-  );
-}
-
-function getChromeDevToolsInstallGuidance(hostName: string) {
-  const normalizedHost = hostName.toLowerCase();
-
-  if (normalizedHost.includes('opencode')) {
-    return [
-      'Install the `chrome-devtools-mcp` server in your OpenCode config:',
-      '',
-      '```json',
-      '{',
-      '  "$schema": "https://opencode.ai/config.json",',
-      '  "mcp": {',
-      '    "chrome-devtools": {',
-      '      "type": "local",',
-      '      "command": ["npx", "-y", "chrome-devtools-mcp@latest"]',
-      '    }',
-      '  }',
-      '}',
-      '```',
-    ].join('\n');
-  }
-
-  if (normalizedHost.includes('claude')) {
-    return ['Install it with:', '', '```bash', 'claude mcp add chrome-devtools --scope user npx chrome-devtools-mcp@latest', '```'].join(
-      '\n',
-    );
-  }
-
-  return [
-    'Install the `chrome-devtools-mcp` server in your MCP client using the setup instructions from:',
-    '',
-    'https://github.com/ChromeDevTools/chrome-devtools-mcp',
-  ].join('\n');
-}
-
-function formatRuntimeRecommendations(
-  recommendations: Array<{ priority: string; message: string; category: string }> | undefined,
-) {
-  const priorityIcon: Record<string, string> = {
-    critical: '🔴',
-    warning: '🟡',
-    info: '💡',
-  };
-
-  const recs = (recommendations ?? []).filter(Boolean);
-  if (recs.length === 0) return '*No follow-up recommendations*';
-
-  return [...recs]
-    .sort((a, b) => {
-      const order: Record<string, number> = { critical: 0, warning: 1, info: 2 };
-      return (order[a.priority] ?? 3) - (order[b.priority] ?? 3);
-    })
-    .map((rec, index) => `${index + 1}. ${priorityIcon[rec.priority] ?? '•'} **[${rec.priority}]** ${rec.message} *(${rec.category})*`)
-    .join('\n');
-}
 
 export default skill({
   name: 'contentful-personalization',
@@ -177,7 +103,7 @@ export default skill({
     }),
     next: ({ response }) => {
       if (response.confidence < 0.6 || response.intent === 'unclear') return 'gather-context';
-      if (response.intent === 'live-debug') return 'live-debug-check-mcp';
+      if (response.intent === 'live-debug') return 'subskill:live-debug';
       if (response.intent === 'reference' && response.topic) return `topic:${response.topic}`;
       if (response.intent === 'reference') return 'pick-topic';
       return `subskill:${response.intent}`;
@@ -217,180 +143,11 @@ export default skill({
       reasoning: 'string',
     }),
     next: ({ response }) => {
-      if (response.intent === 'live-debug') return 'live-debug-check-mcp';
+      if (response.intent === 'live-debug') return 'subskill:live-debug';
       if (response.intent === 'reference' && response.topic) return `topic:${response.topic}`;
       if (response.intent === 'reference') return 'pick-topic';
       return `subskill:${response.intent}`;
     },
-  })
-
-  .step('live-debug-check-mcp', {
-    prompt: ({ host }) => {
-      const matches = getChromeDevToolsToolMatches(host.toolsAvailable);
-
-      return prompt`
-        Determine whether Chrome DevTools MCP is available for this run.
-        Do not ask the user anything. Use only the host tool list below.
-
-        Return:
-        - \`mcpAvailable\` = true if one or more tools clearly belong to Chrome DevTools MCP
-        - \`matchedTools\` = the exact matching tool names shown below
-
-        ## Host
-        ${host.host}
-
-        ## Matching Chrome DevTools MCP tools
-        ${matches.length > 0 ? matches.map((tool) => `- ${tool}`).join('\n') : '(none)'}
-      `;
-    },
-    response: type({
-      mcpAvailable: 'boolean',
-      matchedTools: 'string[]',
-    }),
-    next: ({ response, store }) => {
-      if (!response.mcpAvailable) return 'live-debug-install-mcp';
-      if (store.steps.classify?.requestedUrl || store.steps['gather-context']?.requestedUrl) return 'live-debug-inspect';
-      return 'live-debug-request-url';
-    },
-  })
-
-  .step('live-debug-install-mcp', {
-    prompt: ({ host }) => {
-      const guidance = getChromeDevToolsInstallGuidance(host.host);
-
-      return [
-        prompt`
-          Ask the user to install Chrome DevTools MCP before continuing with live debugging.
-          Explain that runtime inspection is blocked until the MCP server is available in this host.
-          Tell them to rerun the live-debug request after installation because tool availability is fixed for the current run.
-
-          Include this repository link exactly:
-          https://github.com/ChromeDevTools/chrome-devtools-mcp
-
-          Include the host-specific install guidance below exactly as written.
-        `,
-        view('Install Chrome DevTools MCP', guidance),
-      ];
-    },
-    next: terminal,
-  })
-
-  .step('live-debug-request-url', {
-    prompt: act.askUser({
-      type: 'open',
-      question: 'What live URL should I inspect for personalization behavior?',
-    }),
-    response: type({ url: 'string' }),
-    next: 'live-debug-inspect',
-  })
-
-  .step('live-debug-inspect', {
-    prompt: ({ store }) => {
-      const url = getLiveDebugUrl({
-        steps: {
-          classify: store.steps.classify,
-          'gather-context': store.steps['gather-context'],
-          'live-debug-request-url': store.steps['live-debug-request-url'],
-        },
-      });
-
-      return prompt`
-        Use Chrome DevTools MCP to inspect runtime personalization behavior for this live page:
-        ${url}
-
-        ## Required workflow
-        1. Open the page.
-        2. If a cookie consent banner is visible, accept/approve it — personalization and analytics require consent cookies to function fully.
-        3. Wait for it to settle.
-        4. Reload it once so startup requests are visible.
-        5. Inspect console errors and warnings.
-        6. Inspect network traffic, but ONLY requests whose URL contains \`ninetailed.co\` (this includes \`experience.ninetailed.co\` and \`*.insights.ninetailed.co\`).
-        7. If matching requests exist, inspect up to 3 representative requests in detail.
-
-        ## What to report
-        - Whether any meaningful console issues were present
-        - Whether requests to \`ninetailed.co\` were sent (experience API and/or analytics)
-        - Method, status, and a sanitized payload-shape summary for representative requests
-        - Whether the runtime evidence suggests a likely implementation/configuration issue worth escalating into the static doctor flow
-
-        ## Safety rules
-        - Do NOT include cookies, authorization headers, API keys, or full raw payload dumps
-        - Summarize payload shape only
-        - If no matching requests are present, say so explicitly
-
-        Set \`shouldRunDoctor\` to true when the runtime evidence suggests something is off and static code/config diagnosis should continue next.
-        Set it to false only when the runtime behavior looks healthy enough that no static follow-up is needed.
-      `;
-    },
-    response: RuntimeCheckResult,
-    next: 'live-debug-report',
-  })
-
-  .step('live-debug-report', {
-    prompt: ({ store }) => {
-      const result = store.steps['live-debug-inspect'];
-      if (!result) {
-        return prompt`
-          Tell the user the live-debug report could not be rendered because no runtime inspection result was captured.
-          Keep it brief and ask them to rerun the live-debug flow.
-        `;
-      }
-
-      const statusIcon = result.overallStatus === 'pass' ? '✅' : result.overallStatus === 'warn' ? '⚠️' : '❌';
-      const findingsTable = render.table(
-        (result.findings ?? []).map((finding: { item: string; status: string; detail: string }) => ({
-          Check: finding.item,
-          Status: finding.status,
-          Detail: finding.detail,
-        })),
-        { columns: ['Check', 'Status', 'Detail'] },
-      );
-      const requestsTable =
-        (result.requests?.length ?? 0) > 0
-          ? render.table(
-              (result.requests ?? []).map((request: { url: string; method: string; status: number; summary: string }) => ({
-                URL: request.url,
-                Method: request.method,
-                Status: String(request.status),
-                Summary: request.summary,
-              })),
-              { columns: ['URL', 'Method', 'Status', 'Summary'] },
-            )
-          : '*No requests to `ninetailed.co` were detected*';
-
-      const sections = [
-        `# 🌐 Live Debug Report\n`,
-        `## ${statusIcon} Overall: ${result.overallStatus.toUpperCase()}\n`,
-        result.summary,
-        render.section('🖥️ Console Summary', result.consoleSummary),
-        render.section('🌐 ninetailed.co Requests', requestsTable),
-        render.section('🔍 Findings', findingsTable),
-        render.section('💡 Recommendations', formatRuntimeRecommendations(result.recommendations)),
-      ];
-
-      if (result.shouldRunDoctor) {
-        sections.push(
-          render.section(
-            '➡️ Next Step',
-            'The runtime check suggests something is off, so I am continuing into the static doctor flow to inspect the codebase and configuration.',
-          ),
-        );
-      }
-
-      return [
-        prompt`
-          Present the live-debug report below exactly as rendered.
-          If the runtime check looks healthy, keep the wrap-up brief.
-          If the runtime check suggests something is off, tell the user you are continuing into the static doctor flow next.
-        `,
-        view('Live Debug Report', sections.join('\n\n')),
-      ];
-    },
-    next: ({ store }) => (store.steps['live-debug-inspect']?.shouldRunDoctor ? 'subskill:doctor' : 'live-debug-done'),
-  })
-
-  .step('live-debug-done', {
-    next: terminal,
   })
 
   .step('pick-topic', {
@@ -499,16 +256,14 @@ export default skill({
         /ready|readiness|can.*support|prerequisite|pre-check/i.test(''),
     }),
   })
+  .subskill('live-debug', liveDebugSkill, {
+    params: (output) => ({
+      requestedUrl: (output as { requestedUrl?: string })?.requestedUrl,
+    }),
+  })
   .subskill('doctor', doctorSkill, {
-    params: (_output, store) => ({
+    params: () => ({
       userQuery: '',
-      ...(store.steps['live-debug-inspect']
-        ? {
-            triggeredByLiveDebug: true,
-            runtimeUrl: store.steps['live-debug-inspect'].url,
-            runtimeSummary: store.steps['live-debug-inspect'].summary,
-          }
-        : {}),
     }),
   })
   .subskill('develop', developSkill, {
